requirements.txt
groq>=0.9.0
python-dotenv>=1.0.0
fastapi>=0.111.0
uvicorn>=0.29.0
python-multipart>=0.0.9